test fix cyclomatic complexity
